kafka-python/requirements.txt
confluent_kafka==2.2.0
pandas==1.5.2
python-dotenv==1.0.0
